Update Main.java
Changed WINDOW_WIDTH to 1000, and WINDOW_HEIGHT to 800;
Added a global GridPane variable centralUserNtwk for showing the central user's friendship network represented by circles. The GridPane will later be set to the center of the root. The GridPane's width and height will be set as 60% of the WINDOW_WIDTH and 75% of the WINDOW_HEIGHT;
Added the refresh and plot network methods inside IMPORT;
Added a global class UserNode - a texted circle representing a user in a network. Color is predefined as aqua. Constructed by passing a string and a radius;
Added a global public method which returns a GridPane with texted circles representing friends of a central user determined by the SocialNetworkManager mgr.